requirements.txt
matplotlib
tensorflow-addons
tensorflow-model-optimization
tensorflow-probability